+++ src/components/ContentList/__stories__/ContentList.stories.tsx
@@ -1,47 +1,83 @@
 import {Meta, StoryFn} from '@storybook/react';
 
-import {yfmTransform} from '../../../../.storybook/utils';
-import {ContentItemProps, ContentListProps} from '../../../models';
+import {blockTransform} from '../../../../.storybook/utils';
+import {ContentListProps, CustomBlock} from '../../../models';
 import ContentList from '../ContentList';
 
 import data from './data.json';
 
-const transformListItem = (item: ContentItemProps) => ({
-    ...item,
-    text: item?.text && yfmTransform(item.text),
-    title: item?.title && yfmTransform(item.title),
-});
+const BACKGROUND_COLOR = '#313538';
 
 export default {
-    args: {list: data.default.content.map(transformListItem), size: 'l'},
     component: ContentList,
     title: 'Components/ContentList',
 } as Meta;
 
 const DefaultTemplate: StoryFn<ContentListProps> = (args) => (
-    <div style={{paddingBottom: '16px'}}>
-        <ContentList {...args} />
+    <div
+        style={{
+            padding: '16px',
+            backgroundColor: args.theme === 'dark' ? BACKGROUND_COLOR : 'transparent',
+        }}
+    >
+        <ContentList {...(blockTransform(args as unknown as CustomBlock) as ContentListProps)} />
     </div>
 );
 
-const DifferentContentTempalte: StoryFn<ContentListProps> = (args) => (
-    <div style={{paddingBottom: '64px'}}>
-        <h3>Default</h3>
-        <DefaultTemplate {...args} />
-        <h3>Without text</h3>
-        <DefaultTemplate {...args} list={data.withoutText.content.map(transformListItem)} />
-        <h3>Without title</h3>
-        <DefaultTemplate {...args} list={data.withoutTitle.content.map(transformListItem)} />
+const VariousTemplate: StoryFn<Record<number, ContentListProps>> = (args) => (
+    <div style={{display: 'flex', flexDirection: 'column', gap: '16px'}}>
+        {Object.values(args).map((item, index) => (
+            <div
+                key={index}
+                style={{
+                    padding: '8px',
+                    backgroundColor: item.theme === 'dark' ? BACKGROUND_COLOR : 'transparent',
+                }}
+            >
+                <ContentList
+                    {...(blockTransform(item as unknown as CustomBlock) as ContentListProps)}
+                />
+            </div>
+        ))}
     </div>
 );
 
-const DifferentSizeTemplate: StoryFn<ContentListProps> = (args) => (
-    <div>
-        <h1>Size L</h1>
-        <DifferentContentTempalte {...args} size="l" />
-        <h1>Size S</h1>
-        <DifferentContentTempalte {...args} size="s" />
-    </div>
-);
 export const Default = DefaultTemplate.bind({});
-export const Size = DifferentSizeTemplate.bind({});
+export const WithoutText = DefaultTemplate.bind({});
+export const WithoutTitle = DefaultTemplate.bind({});
+export const WithoutIcon = DefaultTemplate.bind({});
+export const Sizes = VariousTemplate.bind({});
+export const Themes = VariousTemplate.bind({});
+
+Default.args = data.default as ContentListProps;
+
+WithoutText.args = data.withoutText as ContentListProps;
+WithoutText.parameters = {
+    controls: {
+        include: Object.keys(data.withoutText),
+    },
+};
+WithoutTitle.args = data.withoutTitle as ContentListProps;
+WithoutTitle.parameters = {
+    controls: {
+        include: Object.keys(data.withoutTitle),
+    },
+};
+WithoutIcon.args = data.withoutIcon as ContentListProps;
+WithoutIcon.parameters = {
+    controls: {
+        include: Object.keys(data.withoutIcon),
+    },
+};
+Themes.args = data.themes as ContentListProps[];
+Themes.parameters = {
+    controls: {
+        include: Object.keys(data.themes),
+    },
+};
+Sizes.args = data.sizes as ContentListProps[];
+Sizes.parameters = {
+    controls: {
+        include: Object.keys(data.sizes),
+    },
+};