+++ src/components/Map/YMap/YandexMap.tsx
@@ -6,7 +6,7 @@ import debounce from 'lodash/debounce';
 import {LocaleContext} from '../../../context/localeContext/localeContext';
 import {MapsContext} from '../../../context/mapsContext/mapsContext';
 import {MobileContext} from '../../../context/mobileContext';
-import {YMapMarkerLabelPrivate, YMapMarkerPrivate, YMapProps} from '../../../models';
+import {YMapMarker, YMapMarkerLabelPrivate, YMapMarkerPrivate, YMapProps} from '../../../models';
 import {block} from '../../../utils';
 import ErrorWrapper from '../../ErrorWrapper/ErrorWrapper';
 import {getMapHeight} from '../helpers';
@@ -26,6 +26,27 @@ const INITIAL_CENTER = [0, 0];
 const BALLOON_DISABLING_MARKER_OPTIONS: YMapMarkerLabelPrivate = {
     cursor: 'drag',
     interactivityModel: 'default#silent',
+};
+
+// Helper function to convert YMapMarker to Schema.org Place object
+const markerToSchemaPlace = (marker: YMapMarker) => {
+    return {
+        '@type': 'Place',
+        address: marker.address
+            ? {
+                  '@type': 'Text',
+                  '@value': marker.address,
+              }
+            : undefined,
+        geo:
+            marker.coordinate && marker.coordinate.length === 2
+                ? {
+                      '@type': 'GeoCoordinates',
+                      latitude: marker.coordinate[0],
+                      longitude: marker.coordinate[1],
+                  }
+                : undefined,
+    };
 };
 
 const YandexMap = (props: YMapProps) => {
@@ -140,6 +161,27 @@ const YandexMap = (props: YMapProps) => {
         }
     }, [ymap, markers, zoom, disableBalloons, areaMargin]);
 
+    const mapMicrodataScript = React.useMemo(() => {
+        if (!markers.length) {
+            return null;
+        }
+
+        const places = markers
+            .filter((marker) => marker.address || marker.coordinate)
+            .map(markerToSchemaPlace);
+
+        if (places.length === 0) {
+            return null;
+        }
+
+        const json = JSON.stringify({
+            '@context': 'https://schema.org',
+            '@graph': places,
+        });
+
+        return <script type="application/ld+json" dangerouslySetInnerHTML={{__html: json}} />;
+    }, [markers]);
+
     if (!markers) return null;
 
     return (
@@ -151,6 +193,7 @@ const YandexMap = (props: YMapProps) => {
             className={b('wrapper')}
         >
             <div className={b('wrapper')}>
+                {mapMicrodataScript}
                 {/* hidden - to show the map after calculating the center */}
                 <div
                     id={containerId}